Changed names of classes and values. Moved logic from main class.

+++ src/main/java/core/basesyntax/service/impl/PurchaseOperation.java
@@ -3,16 +3,18 @@ package core.basesyntax.service.impl;
 import core.basesyntax.dao.RecordDao;
 import core.basesyntax.dao.RecordDaoImpl;
 import core.basesyntax.model.Product;
-import core.basesyntax.record.Record;
-import core.basesyntax.service.DataOperation;
+import core.basesyntax.service.RecordDataManipulation;
 
-public class PurchaseOperation implements DataOperation {
-    private final RecordDao recordDao = new RecordDaoImpl();
+public class PurchaseOperation implements RecordDataManipulation {
+    private final RecordDao recordDao = RecordDaoImpl.getInstance();
 
     @Override
-    public void operate(Record record) {
-        Product product = record.product();
+    public void operate(Product product) {
         Product productFromDB = recordDao.get(product);
+        if (productFromDB == null || productFromDB.getCount() < product.getCount()) {
+            throw new RuntimeException("Not enough products in stock. ProductName="
+                    + product.getName());
+        }
         if (recordDao.remove(product)) {
             recordDao.put(product);
         } else {
